explosion script

+++ Scripts/MISC/DestroyScript.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class DestroyScript : MonoBehaviour
 { 
     // destorys an object after a certain amount of time
-    public int amount = 2;
+    public float amount = 2;
     // Start is called before the first frame update
     void Start()
     {
